<refactoring> UserModel 이름 적용
