+++ back/main_sever/src/auth/dto/auth-credential.dto.ts
@@ -2,7 +2,7 @@ import { PickType } from '@nestjs/mapped-types'
 import { IsEmail, IsString, Matches, MaxLength, MinLength } from 'class-validator'
 import { emailValidationMessage } from 'src/common/validation-message/email-validation.message'
 import { stringValidationMessage } from 'src/common/validation-message/string-validation.message'
-import { UserModel } from 'src/users/entity/user.entity'
+import { UserModel } from 'src/users/entities/user.entity'
 
 // 유저가 로그인할떄 사용할 DTO
 // todo RegisterUserDto 처럼 사용하지 논의필요
