+++ back/main_sever/src/users/user.repository.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common'
 
 import { ConflictException, InternalServerErrorException } from '@nestjs/common'
 import * as bcrypt from 'bcrypt'
-import { UserModel } from './entity/user.entity'
+import { UserModel } from './entities/user.entity'
 import { AuthCredentialsDto, RegisterUserDto } from 'src/auth/dto/auth-credential.dto'
 
 @Injectable()
