+++ back/main_sever/src/users/users.controller.ts
@@ -1,9 +1,22 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common'
+import { Controller, Get, Post, Body, Patch, Param, Delete, UseInterceptors, ClassSerializerInterceptor } from '@nestjs/common'
 import { UsersService } from './users.service'
 import { CreateUserDto } from './dto/create-user.dto'
 import { UpdateUserDto } from './dto/update-user.dto'
 
 @Controller('users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(private readonly usersService: UsersService) {
+  }
+  @Get('/all')
+  // 직렬화를 할 때 데이터 포맷을 변경 (@Exclude()를 이용하면 제거 가능)
+  // @UseInterceptors(ClassSerializerInterceptor) // AppModule에서 전체 적용으로 설정가능
+  /**
+   * serialization -> 직렬화 -> 현재 시스템에서 사용되는 (NestJS) 데이터의 구조를 다른 시스템에서도 쉽게
+   *                          사용 할 수 있는 포맷으로 변환
+   *                        -> class의 object에서 JSON 포맷으로 변환
+   * deserialization -> 역직렬화
+   */
+  getUsers() {
+    return this.usersService.getAllUsers();
+  }
 }
