+++ back/main_sever/src/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { UsersService } from './users.service'
 import { UsersController } from './users.controller'
 import { TypeOrmModule } from '@nestjs/typeorm'
-import { UserModel } from './entity/user.entity'
+import { UserModel } from './entities/user.entity'
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserModel])],
